Hash Table data structure
A hash table is a data structure used to implement an associative
array, that can map keys to values. A hash table uses a hash function
to compute an index into an array of buckets or slots, from which the
correct value can be found. The idea is to access a value in O(1).