Convert code from insertion sort logic to bubble sort
Collection elements should bubble up to the right in a bubble sort.
The compares should be made with all elements to the right.
In an insertion sort elements will sink towards the left and
compares are made with all elements to the left.

Please see step by step compares: https://en.wikipedia.org/wiki/Bubble_sort